좌표 정렬하기 - 나이순 정렬 응용

+++ src/remind/w/PalindromeSum.java
@@ -2,15 +2,15 @@ package remind.w;
 
 import java.util.Scanner;
 
-//3번. 1~n까지의 수를 이진수로 바꿔서 팰린드롬인 수끼리 합을 반환 이야 이걸 못 풀어?
+//3번. 1~n까지의 수를 이진수로 바꿔서 팰린드롬인 수끼리 합을 반환
 public class PalindromeSum {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
         sc.close();
         int sum = 0;
-        for(int i = 1 ; i <= n ; i++) {
-            if(isPalindrome(Integer.toBinaryString(i))){
+        for(long i = 1 ; i <= n ; i++) {
+            if(isPalindrome(Long.toBinaryString(i))){
                 sum += i;
             }
         }
